+++ src/pages/privacy.page.tsx
@@ -1,6 +1,16 @@
+import { AppWrapper } from 'components/common'
+
 import ExternalLink from '../components/ExternalLink'
 
-export default function PrivacyPolicy() {
+export default function PrivacyPolicyPage() {
+  return (
+    <AppWrapper>
+      <PrivacyPolicy />
+    </AppWrapper>
+  )
+}
+
+function PrivacyPolicy() {
   return (
     <div style={{ padding: '40px', maxWidth: '1080px', margin: '0 auto' }}>
       <h1>Privacy Policy</h1>
