+++ src/pages/projects/index.page.tsx
@@ -4,6 +4,7 @@ import { Button } from 'antd'
 import Search from 'antd/lib/input/Search'
 import FeedbackFormButton from 'components/FeedbackFormButton'
 import Loading from 'components/Loading'
+import { AppWrapper } from 'components/common'
 
 import { ProjectCategory } from 'models/project-visibility'
 import React, { useContext, useEffect, useMemo, useRef, useState } from 'react'
@@ -29,13 +30,21 @@ import ProjectsFilterAndSort from './ProjectsFilterAndSort'
 import ArchivedProjectsMessage from './ArchivedProjectsMessage'
 import MyProjects from './MyProjects'
 
+export default function ProjectsPage() {
+  return (
+    <AppWrapper>
+      <Projects />
+    </AppWrapper>
+  )
+}
+
 type OrderByOption = 'createdAt' | 'totalPaid'
 
 const pageSize = 20
 
 const defaultTab: ProjectCategory = 'trending'
 
-export default function Projects() {
+function Projects() {
   const [selectedTab, setSelectedTab] = useState<ProjectCategory>(defaultTab)
 
   // Checks URL to see if tab has been set
