+++ src/pages/v1/create/index.page.tsx
@@ -45,6 +45,7 @@ import {
   uploadProjectMetadata,
 } from 'utils/ipfs'
 import { getTerminalAddress } from 'utils/v1/terminals'
+import { AppWrapper } from 'components/common'
 
 import TicketingForm, {
   TicketingFormFields,
@@ -74,7 +75,15 @@ import { drawerStyle } from 'constants/styles/drawerStyle'
 
 const terminalVersion: V1TerminalVersion = '1.1'
 
-export default function V1Create() {
+export default function V1CreatePage() {
+  return (
+    <AppWrapper>
+      <V1Create />
+    </AppWrapper>
+  )
+}
+
+function V1Create() {
   const router = useRouter()
   const { signerNetwork, userAddress, onSelectWallet } =
     useContext(NetworkContext)
